Replacing EDSDK object with a mostly lint compliant version
Making use of the pre-bound method pattern
