+++ copis/coms/edsdk_object.py
@@ -13,265 +13,280 @@
 # You should have received a copy of the GNU General Public License
 # along with COPISClient.  If not, see <https://www.gnu.org/licenses/>.
 
-"""Implements EDSDK Functionalities"""
+"""Manages COPIS EDSDK Communications"""
 
 import os
 import datetime
 
 from ctypes import c_int, c_uint, c_void_p, sizeof, WINFUNCTYPE
+from dataclasses import dataclass
+from typing import ClassVar
 
 from canon.EDSDKLib import (
     EDSDK, EdsCapacity, EdsPoint, EdsRect, EdsSaveTo,
     EdsShutterButton, EdsSize, Structure)
 
-_edsdk = None
-_console = None
-_camlist = None
-_camref = None
-_camindex = -1
-running = False
+class LocalEDSDK():
+    """Implement EDSDK Functionalities"""
+    def __init__(self) -> None:
+        self._edsdk = None
+        self._console = None
+        self._is_connected = False
+        self._is_waiting_for_image = False
 
-num_cams = 0
+        self._camera = CameraSettings()
+        self._image = ImageSettings()
 
-waiting_for_image = False
-image_filename = None
-image_folder = ''
-image_prefix = 'COPIS'
-image_index = None
 
+    def initialize(self, console) -> None:
+        """Initialize the EDSDK object"""
+        if self._is_connected:
+            return
 
-def initialize(console):
-    if running:
-        return
+        self._console = console
 
-    global _edsdk, _console
-
-    _console = console
-
-    try:
-        _edsdk = EDSDK()
-        _edsdk.EdsInitializeSDK()
-        _update_camera_list()
-
-    except Exception as e:
-        _console.print(f'An exception occurred while initializing Canon API: {e.args[0]}')
-
-
-def connect(index: int = 0) -> bool:
-    """Connect to camera at index, and init it for capturing images.
-
-    Args:
-        index: Defaults to 0.
-
-    Returns:
-        True if successful, False otherwise.
-    """
-    global running, _camref, _camindex
-
-    _update_camera_list()
-
-    # already connected
-    if running and index == _camindex:
-        _console.print(f'Already connected to camera {_camindex}.')
-        return True
-
-    # disconnect from previously connected camera
-    if running:
-        disconnect()
-
-    if num_cams == 0:
-        _console.print('No cameras detected.')
-        return False
-
-    # invalid index
-    if index < 0 or index >= num_cams:
-        _console.print(f'Invalid camera index: {index}.')
-        return False
-
-
-    running = True
-
-    _camindex = index
-    _camref = _edsdk.EdsGetChildAtIndex(_camlist, _camindex)
-    _console.print(f'Connected to camera {_camindex}.')
-
-    _edsdk.EdsOpenSession(_camref)
-    _edsdk.EdsSetPropertyData(_camref, _edsdk.PropID_SaveTo, 0, 4, EdsSaveTo.Host.value)
-    _edsdk.EdsSetCapacity(_camref, EdsCapacity(10000000, 512, 1))
-
-    # set handlers
-    _edsdk.EdsSetObjectEventHandler(_camref, _edsdk.ObjectEvent_All, object_handler, None)
-    _edsdk.EdsSetPropertyEventHandler(_camref, _edsdk.PropertyEvent_All, property_handler, _camref)
-    _edsdk.EdsSetCameraStateEventHandler(_camref, _edsdk.StateEvent_All, state_handler, _camref)
-
-    _edsdk.EdsSetPropertyData(_camref, _edsdk.PropID_Evf_OutputDevice, 0, sizeof(c_uint), _edsdk.EvfOutputDevice_TFT)
-    return True
-
-
-def disconnect() -> bool:
-    """Disconnect from camera.
-
-    Returns:
-        True if successful, False otherwise.
-    """
-    global running, _camref, _camindex
-
-    if not running:
-        _console.print('No cameras currently connected.')
-        return False
-
-    _edsdk.EdsCloseSession(_camref)
-    _console.print(f'Disconnected from camera {_camindex}.')
-
-    running = False
-    _camref = None
-    _camindex = -1
-
-    return True
-
-
-def take_picture() -> bool:
-    """Take picture on connected camera.
-
-    Returns:
-        True if successful, False otherwise.
-    """
-    if not running:
-        _console.print('No cameras currently connected.')
-        return False
-
-    try:
-        global waiting_for_image
-        waiting_for_image = True
-
-        _edsdk.EdsSendCommand(_camref, _edsdk.CameraCommand_PressShutterButton,
-            EdsShutterButton.CameraCommand_ShutterButton_Completely.value)
-
-        _edsdk.EdsSendCommand(_camref, _edsdk.CameraCommand_PressShutterButton,
-            EdsShutterButton.CameraCommand_ShutterButton_OFF.value)
-
-        # while waiting_for_image:
-        #     pythoncom.PumpWaitingMessages()
-
-        return True
-
-    except Exception as e:
-        _console.print(f'An exception occurred while taking a photo with camera {_camindex}: {e.args[0]}')
-        return False
-
-
-def step_focus() -> bool:
-    """TODO
-
-    Returns:
-        True if successful, False otherwise.
-    """
-    return False
-
-
-def start_liveview():
-    """TODO"""
-    return
-
-
-def download_evf_data():
-    """TODO"""
-    return
-
-
-def end_liveview():
-    """TODO"""
-    return
-
-
-def _update_camera_list():
-    """Update camera list and camera count."""
-    global _camlist, num_cams
-
-    _camlist = _edsdk.EdsGetCameraList()
-    num_cams = _edsdk.EdsGetChildCount(_camlist)
-
-
-def terminate():
-    """Terminate EDSDK."""
-    try:
-        global running
-
-        _edsdk.EdsTerminateSDK()
-        running = False
-
-    except Exception as e:
-        _console.print(f'An exception occurred while terminating Canon API: {e.args[0]}')
-
-
-def _generate_file_name():
-    """Sets the filename for an image."""
-    global image_filename, image_folder, image_prefix
-
-    now = datetime.datetime.now().isoformat()[:-7].replace(':', '-')
-    image_filename = os.path.abspath(f'./{image_prefix}_{now}.jpg')
-
-
-def _download_image(image) -> None:
-    """Download image from camera buffer to host computer.
-
-    Args:
-        image: Pointer to the image.
-    """
-    try:
-        _generate_file_name()
-
-        dir_info = _edsdk.EdsGetDirectoryItemInfo(image)
-        stream = _edsdk.EdsCreateFileStream(image_filename, 1, 2)
-
-        _edsdk.EdsDownload(image, dir_info.size, stream)
-        _edsdk.EdsDownloadComplete(image)
-        _edsdk.EdsRelease(stream)
-
-        global waiting_for_image
-        waiting_for_image = False
-
-        _console.print(f'Image saved at {image_filename}.')
-
-    except Exception as e:
-        _console.print(f'An exception occurred while downloading an image: {e.args[0]}')
-
-
-ObjectHandlerType = WINFUNCTYPE(c_int, c_int, c_void_p, c_void_p)
-def _handle_object(event, object, context):
-    """Handles the group of events where request notifications are issued to
-    create, delete or transfer image data stored in a camera or image files on
-    the memory card.
-    """
-    if event == _edsdk.ObjectEvent_DirItemRequestTransfer:
-        _download_image(object)
-    return 0
-object_handler = ObjectHandlerType(_handle_object)
-
-
-PropertyHandlerType = WINFUNCTYPE(c_int, c_int, c_int, c_int, c_void_p)
-def _handle_property(event, property, param, context):
-    """Handles the group of events where notifications are issued regarding
-    changes in the properties of a camera.
-    """
-    return 0
-property_handler = PropertyHandlerType(_handle_property)
-
-
-StateHandlerType = WINFUNCTYPE(c_int, c_int, c_int, c_void_p)
-def _handle_state(event, state, context):
-    """Handles the group of events where notifications are issued regarding
-    changes in the state of a camera, such as activation of a shut-down timer.
-    """
-    if event == _edsdk.StateEvent_WillSoonShutDown:
         try:
-            _edsdk.EdsSendCommand(context, 1, 0)
-        except Exception as e:
-            _console.print(f'An exception occurred while handling the state change event: {e.args[0]}')
-    return 0
-state_handler = StateHandlerType(_handle_state)
+            self._edsdk = EDSDK()
+            self._edsdk.EdsInitializeSDK()
+            self._update_camera_list()
 
+        except Exception as err:
+            self._console.print(
+                f'An exception occurred while initializing Canon API: {err.args[0]}')
+
+
+    def connect(self, index: int = 0) -> bool:
+        """Connect to camera at index, and init it for capturing images.
+
+        Args:
+            index: Defaults to 0.
+
+        Returns:
+            True if successful, False otherwise.
+        """
+        self._update_camera_list()
+
+        cam_count, cam_index, cam_items, cam_ref = self._camera
+
+        # already connected
+        if self._is_connected and index == cam_index:
+            self._console.print(f'Already connected to camera {cam_index}.')
+            return True
+
+        # disconnect from previously connected camera
+        if self._is_connected:
+            self.disconnect()
+
+        if cam_count == 0:
+            self._console.print('No cameras detected.')
+            return False
+
+        # invalid index
+        if index < 0 or index >= cam_count:
+            self._console.print(f'Invalid camera index: {index}.')
+            return False
+
+
+        self._is_connected = True
+
+        self._camera.index = index
+        self._camera.ref = self._edsdk.EdsGetChildAtIndex(cam_items, index)
+        self._console.print(f'Connected to camera {self._camera.index}.')
+
+        *_, cam_ref = self._camera
+
+        self._edsdk.EdsOpenSession(cam_ref)
+        self._edsdk.EdsSetPropertyData(
+            cam_ref, self._edsdk.PropID_SaveTo, 0, 4, EdsSaveTo.Host.value)
+
+        self._edsdk.EdsSetCapacity(cam_ref, EdsCapacity(10000000, 512, 1))
+
+        # set handlers
+        object_prototype = WINFUNCTYPE(c_int, c_int, c_void_p, c_void_p)
+        object_handler = object_prototype(self._handle_object)
+
+        property_prototype = WINFUNCTYPE(c_int, c_int, c_int, c_int, c_void_p)
+        property_handler = property_prototype(self._handle_property)
+
+        state_prototype = WINFUNCTYPE(c_int, c_int, c_int, c_void_p)
+        state_handler = state_prototype(self._handle_state)
+
+        self._edsdk.EdsSetObjectEventHandler(
+            cam_ref, self._edsdk.ObjectEvent_All, object_handler, None)
+
+        self._edsdk.EdsSetPropertyEventHandler(
+            cam_ref, self._edsdk.PropertyEvent_All, property_handler, cam_ref)
+
+        self._edsdk.EdsSetCameraStateEventHandler(
+            cam_ref, self._edsdk.StateEvent_All, state_handler, cam_ref)
+
+        self._edsdk.EdsSetPropertyData(cam_ref, self._edsdk.PropID_Evf_OutputDevice,
+            0, sizeof(c_uint), self._edsdk.EvfOutputDevice_TFT)
+
+        return True
+
+
+    def disconnect(self) -> bool:
+        """Disconnect from camera.
+
+        Returns:
+            True if successful, False otherwise.
+        """
+        if not self._is_connected:
+            self._console.print('No cameras currently connected.')
+            return False
+
+        self._edsdk.EdsCloseSession(self._camera.ref)
+        self._console.print(f'Disconnected from camera {self._camera.index}.')
+
+        self._is_connected = False
+        self._camera.ref = None
+        self._camera.index = -1
+
+        return True
+
+
+    def take_picture(self) -> bool:
+        """Take picture on connected camera.
+
+        Returns:
+            True if successful, False otherwise.
+        """
+        if not self._is_connected:
+            self._console.print('No cameras currently connected.')
+            return False
+
+        try:
+            self._is_waiting_for_image = True
+
+            self._edsdk.EdsSendCommand(self._camera.ref,
+                self._edsdk.CameraCommand_PressShutterButton,
+                EdsShutterButton.CameraCommand_ShutterButton_Completely.value)
+
+            self._edsdk.EdsSendCommand(self._camera.ref,
+                self._edsdk.CameraCommand_PressShutterButton,
+                EdsShutterButton.CameraCommand_ShutterButton_OFF.value)
+
+            # while waiting_for_image:
+            #     pythoncom.PumpWaitingMessages()
+
+            return True
+
+        except Exception as err:
+            self._console.print('An exception occurred while taking a photo with camera '
+                f'{self._camera.index}: {err.args[0]}')
+            return False
+
+
+    def terminate(self):
+        """Terminate EDSDK."""
+        try:
+            self.disconnect()
+            self._edsdk.EdsTerminateSDK()
+
+        except Exception as err:
+            self._console.print(f'An exception occurred while terminating Canon API: {err.args[0]}')
+
+
+    def get_camera_count(self) -> int:
+        """Return camera count"""
+        return self._camera.count
+
+
+    # def step_focus(self) -> bool:
+    #     """TODO
+
+    #     Returns:
+    #         True if successful, False otherwise.
+    #     """
+    #     return False
+
+
+    # def start_liveview(self):
+    #     """TODO"""
+    #     return
+
+
+    # def download_evf_data(self):
+    #     """TODO"""
+    #     return
+
+
+    # def end_liveview(self):
+    #     """TODO"""
+    #     return
+
+
+    def _update_camera_list(self):
+        """Update camera list and camera count."""
+        self._camera.items = self._edsdk.EdsGetCameraList()
+        self._camera.count = self._edsdk.EdsGetChildCount(self._camera.items)
+
+
+    def _generate_file_name(self):
+        """Sets the filename for an image."""
+        now = datetime.datetime.now().isoformat()[:-7].replace(':', '-')
+        self._image.filename = os.path.abspath(f'./{self._image.PREFIX}_{now}.jpg')
+
+
+    def _download_image(self, image) -> None:
+        """Download image from camera buffer to host computer.
+
+        Args:
+            image: Pointer to the image.
+        """
+        try:
+            self._generate_file_name()
+
+            dir_info = self._edsdk.EdsGetDirectoryItemInfo(image)
+            stream = self._edsdk.EdsCreateFileStream(self._image.filename, 1, 2)
+
+            self._edsdk.EdsDownload(image, dir_info.size, stream)
+            self._edsdk.EdsDownloadComplete(image)
+            self._edsdk.EdsRelease(stream)
+
+            self._is_waiting_for_image = False
+
+            self._console.print(f'Image saved at {self._image.filename}.')
+
+        except Exception as err:
+            self._console.print(f'An exception occurred while downloading an image: {err.args[0]}')
+
+
+    def _handle_object(self, evt, obj):
+        """Handles the group of events where request notifications are issued to
+        create, delete or transfer image data stored in a camera or image files on
+        the memory card.
+        """
+        if evt == self._edsdk.ObjectEvent_DirItemRequestTransfer:
+            self._download_image(obj)
+        return 0
+
+
+    def _handle_property(self, evt, prop, param, ctx):
+        """Handles the group of events where notifications are issued regarding
+        changes in the properties of a camera.
+        """
+        return 0
+
+
+    def _handle_state(self, evt, state, ctx):
+        """Handles the group of events where notifications are issued regarding
+        changes in the state of a camera, such as activation of a shut-down timer.
+        """
+        if evt == self._edsdk.StateEvent_WillSoonShutDown:
+            try:
+                self._edsdk.EdsSendCommand(ctx, 1, 0)
+            except Exception as err:
+                self._console.print(
+                    f'An exception occurred while handling the state change event: {err.args[0]}')
+
+        return 0
+
+
+@dataclass
 class EvfDataSet(Structure):
+    """EVF data structure"""
     _fields_ = [
         ('stream', c_void_p),
         ('zoom', c_uint),
@@ -279,3 +294,52 @@ class EvfDataSet(Structure):
         ('imagePosition', EdsPoint),
         ('sizeJpgLarge', EdsSize),
     ]
+
+
+@dataclass
+class CameraSettings():
+    """Data structure to hold the COPIS EDSDK camera settings"""
+    def __init__(self) -> None:
+        self.count = 0
+        self.index = -1
+        self.items = None
+        self.ref = None
+
+
+    def __iter__(self):
+        return iter((
+            self.count,
+            self.index,
+            self.items,
+            self.ref
+        ))
+
+
+@dataclass
+class ImageSettings():
+    """Data structure to hold the COPIS EDSDK image settings"""
+
+    PREFIX: ClassVar[str] = 'COPIS'
+
+    def __init__(self) -> None:
+        self.filename = None
+        self.folder = ''
+        self.index = None
+
+
+    def __iter__(self):
+        return iter((
+            self.filename,
+            self.folder,
+            self.index,
+        ))
+
+
+_instance = LocalEDSDK()
+
+connect = _instance.connect
+disconnect = _instance.disconnect
+get_camera_count = _instance.get_camera_count
+initialize = _instance.initialize
+take_picture = _instance.take_picture
+terminate = _instance.terminate
