+++ copis/gui/main_frame.py
@@ -458,7 +458,7 @@ class MainWindow(wx.Frame):
 
         TODO!
         """
-        if self.core.edsdk.num_cams == 0:
+        if self.core.edsdk.get_camera_count() == 0:
             return
 
         self.panels['evf'] = EvfPanel(self)
